+++ example/lib/pages/money_format_page.dart
@@ -55,15 +55,15 @@ class _MoneyFormatPageState extends State<MoneyFormatPage> {
                     // signed: true,
                   ),
                   inputFormatters: [
-                    // CurrencyInputFormatter(
-                    //   thousandSeparator: ThousandSeparator.Period,
-                    //   mantissaLength: 0,
-                    // )
                     CurrencyInputFormatter(
-                      thousandSeparator: ThousandSeparator.Space,
-                      mantissaLength: 2,
-                      trailingSymbol: "\$",
+                      thousandSeparator: ThousandSeparator.Period,
+                      mantissaLength: 0,
                     )
+                    // CurrencyInputFormatter(
+                    //   thousandSeparator: ThousandSeparator.Space,
+                    //   mantissaLength: 2,
+                    //   trailingSymbol: "\$",
+                    // )
                   ],
                 ),
                 SizedBox(