+++ lib/extensions/exports.dart
@@ -1,0 +1,3 @@
+export 'double_extensions.dart';
+export 'int_extensions.dart';
+export 'string_extensions.dart';